针对vp2的一些测试

+++ UI/FlowView/app/src/main/java/com/arms/flowview/utils/ConverUtils.kt
@@ -6,12 +6,16 @@ import android.util.TypedValue
 /**
  * Created by heyueyang on 2021/11/17
  */
-object ConverUtils{
+object ConverUtils {
 
-    fun dp2px(dp: Int): Int {
+
+    /**
+     * 方法传入[dp]来进行相应的dp转px的操作
+     */
+    fun dp2px(dp: Float): Int {
         return TypedValue.applyDimension(
             TypedValue.COMPLEX_UNIT_DIP,
-            dp.toFloat(),
+            dp,
             Resources.getSystem().displayMetrics
         ).toInt()
     }
